central/requirements.txt
flask==3.1.0
python-telegram-bot[job-queue]==21.9
